Generate email summary table for test-case to infrastructure mapping
